// react-frontend/src/api/models/create-user-profile-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateUserProfileRequest
 */
export interface CreateUserProfileRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateUserProfileRequest
     */
    'credentialNumber': string;
    /**
     * 
     * @type {number}
     * @memberof CreateUserProfileRequest
     */
    'campusId'?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateUserProfileRequest
     */
    'residenceId'?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof CreateUserProfileRequest
     */
    'allergyIds'?: Array<number>;
}