// react-frontend/src/api/models/user-profile-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CompactUserDTO } from './compact-user-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { SelectDTO } from './select-dto';

/**
 * 
 * @export
 * @interface UserProfileDTO
 */
export interface UserProfileDTO {
    /**
     * 
     * @type {number}
     * @memberof UserProfileDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof UserProfileDTO
     */
    'credentialNumber'?: string;
    /**
     * 
     * @type {CompactUserDTO}
     * @memberof UserProfileDTO
     */
    'user'?: CompactUserDTO;
    /**
     * 
     * @type {SelectDTO}
     * @memberof UserProfileDTO
     */
    'campus'?: SelectDTO;
    /**
     * 
     * @type {SelectDTO}
     * @memberof UserProfileDTO
     */
    'residence'?: SelectDTO;
    /**
     * 
     * @type {Array<SelectDTO>}
     * @memberof UserProfileDTO
     */
    'allergies'?: Array<SelectDTO>;
    /**
     * 
     * @type {string}
     * @memberof UserProfileDTO
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfileDTO
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {number}
     * @memberof UserProfileDTO
     */
    'balance'?: number;
}